% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{likelihood}
\usage{
likelihood(rasch, theta)
}
\arguments{
\item{rasch}{a rasch object representing the student in question}

\item{theta}{a numeric representing a quess at the student's ability level}
}
\value{
A number of how likely the proposed theta is to be correct
}
\description{
likelihood
}
\note{
none
}
\examples{

jonah = new("Rasch", name = "Jonah", a = c(5,5,5,10,25), y = c(1,1,0,1,0))
likelihood(jonah, 15)

}
\author{
Jonah Klein-Barton
}
