% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{Log likelihood}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{numeric. The vector of observed data}

\item{lambda}{numeric. The assume value of lambda}
}
\value{
Log likelihood
}
\description{
Calculate the log likelihood for the observed data
}
\examples{
y = c(1,2,3,4)
lambda = 4
logLik(y,lambda)
}
\author{
Berta Diaz

Berta Diaz
}
