\name{midterm-package}
\alias{midterm-package}
\alias{midterm}
\docType{package}
\title{
\packageTitle{midterm}
}
\description{
\packageDescription{midterm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{midterm}
\packageIndices{midterm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{midterm}

Maintainer: \packageMaintainer{midterm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
