% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Maximum likelihood}
\usage{
mle(y)
}
\arguments{
\item{y}{numeric. The vector of observed data}
}
\value{

}
\description{
Calculates the maximum likelihood estimator for lambda
}
\examples{
y=c(1,2,3,4)
mle(y)
}
\author{
Berta Diaz
}
