% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simp.R
\name{plot_simp}
\alias{plot_simp}
\alias{plot_simp,}
\alias{simpson-method}
\title{Plots the graph of a function using simpson sums}
\usage{
plot_simp(simp)
}
\arguments{
\item{simp}{An object of the class 'simpson'}
}
\value{
The plot of the results of the calculation
}
\description{
Plots the results of simpson sums
}
\author{
Joseph Ludmir
}

