% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trap.R
\name{plot_trap}
\alias{plot_trap}
\alias{plot_trap,}
\alias{trapezoid-method}
\title{Plots the graph of a function using trapezoidal sums}
\usage{
plot_trap(trap)
}
\arguments{
\item{trap}{An object of the class 'trapezoid'}
}
\value{
The plot of the results of the calculation
}
\description{
Plots the results of trapezoidal sums
}
\author{
Joseph Ludmir
}

