% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{y}}{numeric.}

\item{\code{mle}}{numeric.}

\item{\code{ll}}{numeric.}

\item{\code{se}}{numeric.}

\item{\code{SEtype}}{character.}
}}

