% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,Rasch-method}
\alias{print,Rasch-method}
\title{print}
\usage{
\S4method{print}{Rasch}(x)
}
\arguments{
\item{x}{a rasch object representing the student in question}
}
\value{
Prints the students name and the output of their respective EAP predicted ability (theta)
}
\description{
print
}
\note{
none
}
\examples{

jonah = new("Rasch", name = "Jonah", a = c(5,5,5,10,25), y = c(1,1,0,1,0))
print(jonah)

}
\author{
Jonah Klein-Barton
}
