% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_simp.R
\name{print_simp}
\alias{print_simp}
\alias{print_simp,}
\alias{simpson-method}
\title{Print methods for simpson sums}
\usage{
print_simp(simp)
}
\arguments{
\item{simp}{An object of the class 'simpson'}
}
\value{
The result of the calculation
}
\description{
Prints the results of simpson sums
}
\author{
Joseph Ludmir
}

