% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_trap.R
\name{print_trap}
\alias{print_trap}
\alias{print_trap,}
\alias{trapezoid-method}
\title{Print methods for trapezoidal sums}
\usage{
print_trap(trap)
}
\arguments{
\item{trap}{An object of the class 'trapezoid'}
}
\value{
The result of the calculation
}
\description{
Prints the results of trapezoidal sums
}
\author{
Joseph Ludmir
}

