% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{prior}
\usage{
prior(theta)
}
\arguments{
\item{theta}{a numeric representing a quess at the student's ability level}
}
\value{
The height of the normal curve where x = theta, the mean is zero, and the standard deviation is three
}
\description{
prior
}
\note{
none
}
\examples{

theta = 15
prior(15)

}
\author{
Jonah Klein-Barton
}
