% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{probability}
\usage{
probability(raschObj, theta)
}
\arguments{
\item{raschObj}{a rasch object representing the student in question}

\item{theta}{a numeric representing a quess at the student's ability level}
}
\value{
A list containing P and PQ, where P represents the estimate that a student of the estimated ability would get the question correct, and PQ displays P if the student got the question and 1-P if they did not
}
\description{
probability
}
\note{
none
}
\examples{

jonah = new("Rasch", name = "Jonah", a = c(5,5,5,10,25), y = c(1,1,0,1,0))
probability(jonah, 15)

}
\author{
Jonah Klein-Barton
}
