% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson.R
\docType{class}
\name{simpson-class}
\alias{initialize,simpson-method}
\alias{simpson-class}
\title{A simpson object.}
\description{
Object of class \code{simpson} are created by the \code{integrateIt} function
which is created by \code{x} and \code{y} and \code{a} and \code{b} and \code{choose_calc}
and the result is printed by \code{print_trap}
}
\details{
An object of the class `simpson' has the following slots:
\itemize{
\item \code{x} The x-values of the function.
\item \code{y} The y-values of the function
\item \code{a} The starting point on the x-axis of integration.
\item \code{b} The ending point on the x-axis of integration.
\item \code{calc} The resulting Simpson sum by integration.
}
}
\author{
Joseph Ludmir: \email{jludmir@wustl.edu}
}

