% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid.R
\docType{class}
\name{trapezoid-class}
\alias{gettrapezoid}
\alias{initialize,trapezoid-method}
\alias{trapezoid-class}
\title{A trapezoidal object.}
\usage{
gettrapezoid(object = "trapezoid")
}
\description{
Object of class \code{trapezoid} are created by the \code{integrateIt} function
which is created by \code{x} and \code{y} and \code{a} and \code{b} and \code{choose_calc}
}
\details{
An object of the class `Trapezoid' has the following slots:
\itemize{
\item \code{x} The x-values of the function.
\item \code{y} The y-values of the function
\item \code{calc} The resulting Trapezoidal sum by integration.
}
}
\author{
Joseph Ludmir: \email{jludmir@wustl.edu}
}

