% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_simp.R, R/print_trap.R
\name{print_simp}
\alias{print,simpson-method}
\alias{print,trapezoid-method}
\alias{print_simp}
\alias{print_trap}
\title{Print methods for simpson sums}
\usage{
print_simp(simp)

print_trap(trap)
}
\arguments{
\item{simp}{An object of the class 'simpson'}

\item{trap}{An object of the class 'trapezoid'}
}
\value{
The result of the calculation

The result of the calculation
}
\description{
Prints the results of simpson sums

Prints the results of trapezoidal sums
}
\author{
Joseph Ludmir

Joseph Ludmir
}

