% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{Probability}
\alias{Probability}
\title{Runs Probability function}
\usage{
Probability(raschObj, theta)
}
\arguments{
\item{raschObj}{a object of class Rasch}

\item{theta}{numeric designating difficulty}
}
\value{
two lists. one of P.i.j. One of PQ
}
\description{
Runs Probability function
}
\note{
It works with a for loop
}
\examples{

}
\author{
Jack Ploshnick
}
