% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch Class.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\title{Creates Rasch Class}
\arguments{
\item{name}{a character}

\item{a_value}{difficulty}

\item{y_j_value}{dummy for correct answer}
}
\value{
Class of Rasch
}
\description{
Creates Rasch Class
}
\note{
Makes the Class
}
\examples{

}
\author{
Jack Ploshnick
}
